removing more of the old locking infrastructure and cleaning up a bit
git-svn-id: https://qthreads.googlecode.com/svn/trunk@3459 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ include/qt_debug.h
@@ -318,7 +318,7 @@ static QINLINE void qthread_debug(int         level,
                         if (str == NULL) {
                             qassert(WRITE(2, "(null)", 6), 6);
                         } else {
-                            qassert(WRITE(2, str, strlen(str)), strlen(str));
+                            qassert(WRITE(2, str, strlen(str)), (ssize_t)strlen(str));
                         }
                         break;
                     }
